requirements.txt
azure-mgmt-resource>=20.0.0
azure-identity>=1.5.0